requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.3
django-cors-headers==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
